Ranking aanmaken +wegschrijven db

+++ IP3_8IEN/Domain/Dashboard/GraphData.cs
@@ -6,7 +6,7 @@ namespace IP3_8IEN.BL.Domain.Dashboard
 {
     public class GraphData
     {
-        public GraphData(string label, int value)
+        public GraphData(string label, double value)
         {
             this.label = label;
             this.value = value;
@@ -19,7 +19,7 @@ namespace IP3_8IEN.BL.Domain.Dashboard
         [Key]
         public int GraphDataId { get; set; }
         public string label { get; set; }
-        public int value { get; set; }
+        public double value { get; set; }
 
         public DashItem DashItem { get; set; }
     }
